feat(dashboard-refresh): fix due to design change
